requirements.txt
asgiref==3.3.1
Django==3.1.6
django-appconf==1.0.4
django-imagekit==4.0.2
freeze==3.0
importlib-metadata==3.6.0
Markdown==3.3.3
mysqlclient==2.0.3
pilkit==2.0
Pillow==8.1.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
